Fixed color size

+++ src/devices/base.h
@@ -12,7 +12,7 @@ class PaperdinkDeviceBaseClass
 {
     public:
         uint8_t has_color = false;
-        uint8_t color = GxEPD_BLACK;
+        int color = GxEPD_BLACK;
 
         int8_t begin();
         virtual int8_t enable_display();
