+++ src/devices/merlot.h
@@ -8,7 +8,7 @@ class Paperdink_Merlot : public PaperdinkDeviceBaseClass
 {
     public:
         uint8_t has_color = true;
-        uint8_t color = GxEPD_RED;
+        int color = GxEPD_RED;
 
         GxEPD2_3C<GxEPD2_420c_Z21, GxEPD2_420c_Z21::HEIGHT> epd = GxEPD2_420c_Z21(EPD_CS, EPD_DC, EPD_RST, EPD_BUSY);
         int8_t enable_display() override;
